feat: implement use MongoDB database in board and task methods

+++ src/resources/users/user.DB.repository.js
@@ -1,4 +1,5 @@
 const { User } = require('./user.model');
+const taskService = require('../tasks/task.service');
 
 const {
   // ValidationError,
@@ -23,7 +24,12 @@ const update = async (id, user) => {
 };
 
 const remove = async id => {
-  await User.deleteOne({ _id: id });
+  const user = await User.findById(id);
+  if (!user) {
+    throw new NotFoundError(`Couldn't find a user with id: ${id}`);
+  }
+  const handledTasks = await taskService.deleteUserTasks(id);
+  const deletedUser = await User.deleteOne({ _id: id });
+  return { handledTasks, deletedUser };
 };
-
 module.exports = { getAll, save, get, remove, update };
